+++ src/App.tsx
@@ -3,6 +3,7 @@ import Layout from './components/layout/Layout'
 import HomePage from './pages/HomePage'
 import NumbersPage from './pages/NumbersPage'
 import DatesPage from './pages/DatesPage'
+import CurrencyPage from './pages/CurrencyPage'
 
 function App() {
   return (
@@ -12,6 +13,7 @@ function App() {
           <Route index element={<HomePage />} />
           <Route path="numbers" element={<NumbersPage />} />
           <Route path="dates" element={<DatesPage />} />
+          <Route path="currency" element={<CurrencyPage />} />
         </Route>
       </Routes>
     </BrowserRouter>